requirements.txt
alembic==1.3.3
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
coverage==5.0.3
coveralls==1.10.0
docopt==0.6.2
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.0
Mako==1.1.1
MarkupSafe==1.1.1
more-itertools==8.1.0
packaging==20.1
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.1
pyparsing==2.4.6
pytest==5.3.4
pytest-cov==2.8.1
python-dateutil==2.8.1
python-dotenv==0.10.5
python-editor==1.0.4
requests==2.22.0
six==1.14.0
SQLAlchemy==1.3.13
SQLAlchemy-Utils==0.36.1
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==0.16.1
